some cleanups
